chore: refactors code to folders
