+++ .vscode/tasks.json
@@ -1,0 +1,31 @@
+{
+	"version": "2.0.0",
+	"tasks": [
+		{
+			"type": "typescript",
+			"tsconfig": "tsconfig.json",
+			"problemMatcher": ["$tsc"],
+			"group": {
+				"kind": "build",
+				"isDefault": true
+			},
+			"label": "tsc: build - tsconfig.json"
+		},
+		{
+			"type": "shell",
+			"command": "bun",
+			"args": ["run", "index.ts"],
+			"group": "test",
+			"label": "Run index.ts with Bun",
+			"problemMatcher": []
+		},
+		{
+			"type": "shell",
+			"command": "bun",
+			"args": ["run", "lint"],
+			"group": "test",
+			"label": "Run lint with Bun",
+			"problemMatcher": []
+		}
+	]
+}
